Separate the back and front end apps.
This makes the repo a monorepo.

The existing settings have all moved to the back-end folder.

The listen react app moves to the front-end folder.

In addition, we are now using dotenv explicitly in the settings module so we don't need so many config variables.

This commit just gets us to a local working build.