Guard for race condition on client posts.
Because the client is receiving multiple APNS notifications on a different thread than it's confirming receipt of those notifications, there can be race conditions between receipt of new secrets and posting of old secrets.  The fix is to remember the last secret we saw in an APNS posting that wasn't the secret we had given the client, and if the client posts it again after confirming receipt of a new secret then we recognize that the client is sending us out of date data.

Fixes #2.