Revamp the front end.
1. Use Material UI
2. Add disconnect button.
3. Always make user agree to advisory before listening.

NOTE: The always requiring user to agree was originally there to make sure we could do an unload hook.  But unload hooks don't seem to work.  We are leaving it there to make it easy to change your client name and to ensure people don't just "connect and walk away".