Overlap transcription on server transfer.
Now, when one server is suspending a transcript and another is resuming it, they overlap for several seconds to make sure no packets are lost.  In the overlap period, they each save the ID of the transcribed packet so the transcription process can know which packets are duplicates (saved by both servers).