Remove overly broad transcript termination.
There was an assumption that a client and profile could only be involved in one conversation at a time. That's not true since multiple windows were introduced.

What is true is that a single client/profile/conversation pair can only have one content ID at a time (i.e., only one session for the conversation).